alioth: Disable NFC and eSE services on Mi 11X
While NFC controller exists on aliothin, the coil pad is removed from the
final product effecting making it useless to be enabled on the device.

Use ODM SKU to ensure that it's only enabled on Redmi K40 and POCO F3, which
has ro.boot.product.hardware.sku set to `pro'.

This is done based on comparing with MIUI software changes done for aliothin.

Signed-off-by: Albert I <kras@raphielgang.org>