alioth: overlay: Configure display cutout and status bar heights
A circle cutout is drawn based on approximation that's taken from stock MIUI,
but slightly edited to leave as minimum space below the camera cutout without
leaving any visible jagged lines surounding the camera.

Status bar height is taken from stock as it is, and is already perfect so no
modifications needed for this one.

Signed-off-by: Albert I <kras@raphielgang.org>